requirements.txt
autopep8==1.5.4
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
filetype==1.0.7
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.7
jupyter-core==4.7.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.9
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
Pygments==2.7.3
python-dateutil==2.8.1
python-magic==0.4.18
python-magic-bin==0.4.14
pywin32==300
pyzmq==20.0.0
rsa==4.6
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
